requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.37.1
h5py==3.1.0
importlib-metadata==4.0.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
pkg-resources==0.0.0
protobuf==3.16.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
